refactored ImageProcessor into a new package Filter. This package contains FilterHelper, CircleFilter, RectangleFilter and Benchmark classes. Also adjusted GUIController and DRoneController so that they comply with these changes. Corrected access control of methods to public.
