+++ autonomous-drone/src/main/java/dk/localghost/hold17/autonomous_drone/controller/DroneController.java
@@ -1,7 +1,7 @@
 package dk.localghost.hold17.autonomous_drone.controller;
 
 import dk.localghost.hold17.autonomous_drone.opencv_processing.Direction;
-import dk.localghost.hold17.autonomous_drone.opencv_processing.ImageProcessor;
+import dk.localghost.hold17.autonomous_drone.opencv_processing.filter.RectangleFilter;
 import dk.localghost.hold17.base.IARDrone;
 import dk.localghost.hold17.base.command.CommandManager;
 import dk.localghost.hold17.base.command.LEDAnimation;
@@ -197,10 +197,10 @@ public class DroneController {
     }
 
     private Direction getPaperDirection() {
-        ImageProcessor imageProcessor = new ImageProcessor();
+        RectangleFilter rectangleFilter = new RectangleFilter();
 
-        imageProcessor.findBiggestQRCode(imageProcessor.filterImage(droneCamera));
-        return imageProcessor.findPaperPosition(imageProcessor.getBiggestQRCode());
+        rectangleFilter.findBiggestQRCode(rectangleFilter.filterImage(droneCamera));
+        return rectangleFilter.findPaperPosition(rectangleFilter.getBiggestQRCode());
     }
 
     public void LEDSuccess() {
